Migrate `tests/unit/modules/test_archive.py` to PyTest
